hotfix(finance): romper cadena env-leak en /admin/finanzas/resumen
Causa raíz exacta:
- src/lib/queries/financeDashboard/financeResumen.ts línea 1 tenía
  'server-only'; (string literal — NO-OP, no protege nada). El import
  real es 'import "server-only";'.
- FinanceMonthlyControl.tsx (client) importaba buildContextualText y
  EXPENSE_SUBTYPE_LABELS de ese archivo.
- financeResumen.ts importa @/lib/db.
- @/lib/db línea 6 accede a env.DATABASE_URL a module-init.
- Resultado en cliente: t3-env proxy lanza
  "Attempted to access a server-side environment variable on the client".

Por qué se exponía ahora: cadena pre-existente. PR #133 (PAYROLL_OCR_ENABLED
con .transform) cambió la firma del schema t3-env y Turbopack dejó de
tree-shaker la rama db de financeResumen para el bundle cliente. La cadena
client → financeResumen → db → env quedó en el bundle.

Fix mínimo (sin tocar nada de OCR/Blob/RBAC/schema/queries de DB):
- Nuevo src/lib/queries/financeDashboard/financeResumen.shared.ts con
  tipos, constantes y helpers puros. SIN imports de @/lib/db ni @/lib/env.
- financeResumen.ts:
  - 'server-only' (string literal) → import 'server-only' (real).
  - Re-exporta del shared para mantener compat con server callers.
  - Importa helpers puros desde el shared para uso interno.
- FinanceMonthlyControl.tsx y FinanceResumenBlocks.tsx importan de
  financeResumen.shared.ts directamente.
- Test finance-resumen.test.ts ajustado a .shared.ts.

Regresión-shield:
- Nuevo test estático src/__tests__/server/client-server-boundary.test.ts
  que escanea todo src/ y falla si cualquier 'use client' importa
  @/lib/env, @/lib/db, @/lib/auth o @/lib/auth-guard.
- Excepción documentada: ConsentedScripts.tsx solo accede a
  env.NEXT_PUBLIC_GTM_ID (client var, permitido por t3-env proxy).

Validaciones:
- npx tsc --noEmit ✅
- npm run lint ✅
- npm test ✅ — 93 suites / 1597 tests

No tocado: OCR, Blob, RBAC, Finanzas UI (salvo imports de los 2 client
components), schema, queries DB, env.ts, invoices, deploy.
